Add views and scripts for order delivery and voucher management
- Created a new view for received orders with filtering options and order details.
- Implemented JavaScript functions for handling order actions such as search, cancel, and complete.
- Added a new view for managing user vouchers with filtering and detailed information.
- Developed corresponding JavaScript functions for voucher actions including searching and copying voucher codes.
- Introduced a new API integration for fetching shippers and handling order assignments.
- Enhanced user experience with toast notifications for success, warning, and error messages.

+++ Assignment/Models/Users.cs
@@ -28,10 +28,10 @@ namespace Assignment.Models
         public string Role { get; set; } = "Customer";
         [Required]
         [DefaultValue(0)]
-        public long TotalAccumulatedPoints { get; set; } = 0;
+        public double TotalAccumulatedPoints { get; set; } = 0;
         [Required]
         [DefaultValue(0)]
-        public long AccumulatedPoints { get; set; } = 0;
+        public double AccumulatedPoints { get; set; } = 0;
         [Required]
         [DefaultValue(UserRankEnum.None)]
         public UserRankEnum Rank { get; set; } = UserRankEnum.None;
